Functional programming - Function with lambda function as parameter
The idea of this exercise was to create a function capable of receiving a lambda
function as part of the parameter. In this case, it  was used a Predicate to test a
given condition, allowing to build a generic filtering function that iterates a list
of products and sums their values based on a test that it receives (i.e. the filter)
as a parameter