Functional programming - Final exercise
Given a .csv file containing employees (name, email and salary).
read their data. Then, show in alphabetical order the email of
employees whose salary are superior to a certain value provided
by the user. Also show the sum of the salaries of employees whose
names starts with the letter 'M'